Split rules that can be used to cross compile
