+++ scalafix/input/src/main/scala/fix/CopyToBufferSrc.scala
@@ -1,13 +1,11 @@
 /*
-rule = "scala:fix.Stable"
+rule = "scala:fix.CrossCompat"
  */
 package fix
 
 import scala.collection.mutable
 
 class CopyToBufferSrc(xs: List[Int], b: mutable.Buffer[Int]) {
-
   xs.copyToBuffer(b)
   (xs ++ xs).copyToBuffer(b)
-
 }
