+++ scalafix/output212/src/main/scala/fix/CopyToBufferSrc.scala
@@ -6,8 +6,6 @@ package fix
 import scala.collection.mutable
 
 class CopyToBufferSrc(xs: List[Int], b: mutable.Buffer[Int]) {
-
   b ++= xs
   b ++= xs ++ xs
-
 }
