+++ scalafix/output212/src/main/scala/fix/FoldSrc.scala
