+++ scalafix/output212/src/main/scala/fix/LinearSeqSrc.scala
