+++ scalafix/output213/src/main/scala/fix/UnstableStreamToLazyListSrc.scala
