+++ scalafix/rules/src/main/scala/fix/CrossCompat.scala
@@ -1,0 +1,5 @@
+package fix
+
+import scalafix._
+
+case class CrossCompat(index: SemanticdbIndex) extends SemanticRule(index, "CrossCompat") with Stable212Base
