+++ scalafix/rules/src/main/scala/fix/Experimental.scala
@@ -5,53 +5,21 @@ import scalafix.syntax._
 import scalafix.util._
 import scala.meta._
 
+// 2.12 Cross-Compatible
 case class Experimental(index: SemanticdbIndex) extends SemanticRule(index, "Experimental") {
-  // WARNING: TOTAL HACK
-  // this is only to unblock us until Term.tpe is available: https://github.com/scalameta/scalameta/issues/1212
-  // if we have a simple identifier, we can look at his definition at query it's type
-  // this should be improved in future version of scalameta
-  object TypeMatcher {
-    def apply(symbols: Symbol*)(implicit index: SemanticdbIndex): TypeMatcher =
-      new TypeMatcher(symbols: _*)(index)
-  }
 
-  final class TypeMatcher(symbols: Symbol*)(implicit index: SemanticdbIndex) {
-    def unapply(tree: Tree): Boolean = {
-      index.denotation(tree)
-           .exists(_.names.headOption.exists(n => symbols.exists(_ == n.symbol)))
-    }
-  }
-
-  val CollectionMap: TypeMatcher = TypeMatcher(
+  val CollectionMap = TypeMatcher(
     Symbol("_root_.scala.collection.immutable.Map#"),
     Symbol("_root_.scala.collection.mutable.Map#"),
     Symbol("_root_.scala.Predef.Map#")
   )
+  val CollectionSet = TypeMatcher(Symbol("_root_.scala.collection.Set#"))
 
-  val CollectionSet: TypeMatcher = TypeMatcher(Symbol("_root_.scala.collection.Set#"))
-
-  val mapZip = 
-    SymbolMatcher.exact(
-      Symbol("_root_.scala.collection.IterableLike#zip(Lscala/collection/GenIterable;Lscala/collection/generic/CanBuildFrom;)Ljava/lang/Object;.")
-    )
-
-  val mapPlus = 
-    SymbolMatcher.exact(
-      Symbol("_root_.scala.collection.MapLike#`+`(Lscala/Tuple2;)Lscala/collection/Map;.")
-    )
-
-  val setPlus = 
-    SymbolMatcher.exact(
-      Symbol("_root_.scala.collection.SetLike#`+`(Ljava/lang/Object;)Lscala/collection/Set;.")
-    )
-
-  val setMinus = 
-    SymbolMatcher.exact(
-      Symbol("_root_.scala.collection.SetLike#`-`(Ljava/lang/Object;)Lscala/collection/Set;.")
-    )
-
-  def startsWithParens(tree: Tree): Boolean = 
-    tree.tokens.headOption.map(_.is[Token.LeftParen]).getOrElse(false)
+  // == Symbols ==
+  val mapZip   = exact("_root_.scala.collection.IterableLike#zip(Lscala/collection/GenIterable;Lscala/collection/generic/CanBuildFrom;)Ljava/lang/Object;.")
+  val mapPlus  = exact("_root_.scala.collection.MapLike#`+`(Lscala/Tuple2;)Lscala/collection/Map;.")
+  val setPlus  = exact("_root_.scala.collection.SetLike#`+`(Ljava/lang/Object;)Lscala/collection/Set;.")
+  val setMinus = exact("_root_.scala.collection.SetLike#`-`(Ljava/lang/Object;)Lscala/collection/Set;.")
 
   def replaceMapZip(ctx: RuleCtx): Patch = {
     ctx.tree.collect {
@@ -65,7 +33,7 @@ case class Experimental(index: SemanticdbIndex) extends SemanticRule(index, "Exp
       val col = "_root_.scala.collection." + col0
       val callSite =
         if (startsWithParens(rhs)) {
-          ctx.addLeft(rhs, col)          
+          ctx.addLeft(rhs, col)
         }
         else {
           ctx.addLeft(rhs, col + "(") +
@@ -78,7 +46,7 @@ case class Experimental(index: SemanticdbIndex) extends SemanticRule(index, "Exp
     ctx.tree.collect {
       case Term.ApplyInfix(CollectionSet(), op @ setPlus(_), Nil, List(rhs)) =>
         rewriteOp(op, rhs, "+", "Set")
-        
+
       case Term.ApplyInfix(CollectionSet(), op @ setMinus(_), Nil, List(rhs)) =>
         rewriteOp(op, rhs, "-", "Set")
