+++ scalafix/tests/src/test/scala/fix/Collectionstrawman_Tests.scala
@@ -9,8 +9,9 @@ class Collectionstrawman_Tests
     SemanticdbIndex.load(Classpath(AbsolutePath(BuildInfo.inputClassdirectory))),
     AbsolutePath(BuildInfo.inputSourceroot),
     Seq(
-      AbsolutePath(BuildInfo.outputSourceroot), 
-      AbsolutePath(BuildInfo.outputFailureSourceroot)
+      AbsolutePath(BuildInfo.output212Sourceroot),
+      AbsolutePath(BuildInfo.output213Sourceroot),
+      AbsolutePath(BuildInfo.output213FailureSourceroot)
     )
   ) {
   runAllTests()
